Reimplement using reactive-streams
