+++ rx-java/src/main/java/io/vertx/rx/java/RxHelper.java
@@ -4,9 +4,14 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
+import io.vertx.core.streams.Pump;
 import io.vertx.core.streams.ReadStream;
+import io.vertx.ext.reactivestreams.ReactiveWriteStream;
+import org.reactivestreams.Publisher;
+import org.reactivestreams.Subscription;
 import rx.Observable;
 import rx.Observer;
+import rx.RxReactiveStreams;
 import rx.Scheduler;
 import rx.Subscriber;
 import rx.functions.Action0;
@@ -30,7 +35,56 @@ public class RxHelper {
    * @return the adapted observable
    */
   public static <T> Observable<T> toObservable(ReadStream<T> stream, Vertx vertx) {
-    return Observable.create(new ReadStreamAdapter<>(stream));
+    ReactiveWriteStream<T> ws = ReactiveWriteStream.writeStream(vertx);
+    Observable<T> obs = RxReactiveStreams.toObservable(new Publisher<T>() {
+      @Override
+      public void subscribe(org.reactivestreams.Subscriber<? super T> subscriber) {
+        org.reactivestreams.Subscriber<T> subscriberProxy = new org.reactivestreams.Subscriber<T>() {
+          Pump pump = Pump.pump(stream, ws);
+          @Override
+          public void onSubscribe(Subscription subscription) {
+            Subscription subscriptionProxy = new Subscription() {
+              @Override
+              public void request(long n) {
+                subscription.request(n);
+              }
+              @Override
+              public void cancel() {
+                vertx.runOnContext(v -> {
+                  subscriber.onComplete();
+                });
+                pump.stop();
+                subscription.cancel();
+//                ws.close(); ?
+              }
+            };
+            stream.endHandler(done -> {
+              subscriptionProxy.cancel();
+            });
+            stream.exceptionHandler(this::onError);
+            pump.start();
+            subscriber.onSubscribe(subscriptionProxy);
+          }
+
+          @Override
+          public void onNext(T event) {
+            subscriber.onNext(event);
+          }
+
+          @Override
+          public void onError(Throwable t) {
+            subscriber.onError(t);
+          }
+
+          @Override
+          public void onComplete() {
+            subscriber.onComplete();
+          }
+        };
+        ws.subscribe(subscriberProxy);
+      }
+    });
+    return obs;
   }
 
   /**
